// libs/core/infrastructure/config/types/general/workflow-queue.type.ts
export type WorkflowQueueConfig = {
    WORKFLOW_QUEUE_WORKER_PREFETCH: number;
    WORKFLOW_QUEUE_WORKER_MAX_RETRIES: number;
    WORKFLOW_QUEUE_WORKER_RETRY_DELAY_MS: number;
    WORKFLOW_QUEUE_CIRCUIT_BREAKER_FAILURE_THRESHOLD: number;
    WORKFLOW_QUEUE_CIRCUIT_BREAKER_TIMEOUT_MS: number;
    WORKFLOW_QUEUE_AUTO_SCALE_ENABLED: boolean;
    WORKFLOW_QUEUE_AUTO_SCALE_MIN_WORKERS: number;
    WORKFLOW_QUEUE_AUTO_SCALE_MAX_WORKERS: number;
    WORKFLOW_QUEUE_AUTO_SCALE_QUEUE_THRESHOLD: number;
};